requirements.txt
Pillow==8.3.2  # https://github.com/python-pillow/Pillow
psycopg2==2.9.1

#Django
#-------------------------------------
django==3.2.7
django-environ==0.6.0  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-filter==2.4.0  # https://github.com/carltongibson/django-filter
django-extensions==3.1.3 # https://github.com/django-extensions/django-extensions
djangorestframework==3.12.4
dj-database-url==0.5.0
ipython==7.27.0
django-rest-swagger==2.2.0
drf-spectacular==0.17.0
django-cors-headers==3.8.0
